+++ NEWS.md
@@ -1,11 +1,13 @@
 #Revisions and Change Log
 
+### v0.4.7
+* Added shinydashboard plus functionality for a right-hand sidebar as an option
+* Updated documentation and examples for shinydashboardplus functionality
+* Added a preference to turn off the reset application button
+
 ### v0.4.6
 * Bugfix - hide downloadable table button if there are no download functions defined
 * Updated tests to be compatable with the next release of htmltools (0.4, schloerke)
-
-### v0.4.5
-* Bugfix - downloadable table button was not appearing when created in a reactive block
 
 ### v0.4.5
 * Bugfix - downloadable table button was not appearing when created in a reactive block